api-python/requirements.txt
# Azure Functions for Python
azure-functions>=1.11.0

# Azure Face API SDK (replaces dlib and face-recognition)
azure-cognitiveservices-vision-face>=0.6.0
msrest>=0.6.21

# Core dependencies
numpy>=1.21.0,<2.0.0  # Pin numpy to 1.x for compatibility
Pillow>=9.0.0

# Azure Storage SDK
azure-storage-blob>=12.14.0

# Database driver
pyodbc>=4.0.35

# Additional dependency for werkzeug (already specified but making explicit)
werkzeug~=3.1.3
